+++ memory/ft_realloc.c
@@ -6,14 +6,14 @@
 /*   By: snicolet <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/08/25 12:07:53 by snicolet          #+#    #+#             */
-/*   Updated: 2015/12/02 12:48:31 by snicolet         ###   ########.fr       */
+/*   Updated: 2016/10/12 17:38:03 by snicolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-void	*ft_realloc(void *ptr, unsigned int size, unsigned int new_size)
+void	*ft_realloc(void *ptr, size_t size, size_t new_size)
 {
 	void	*dest;
